Mv SMVGEAR/KPP funcs. to their own files + housekeeping
Added more reference species for processing code
